Add user accounts and authentication

+++ database/index.js
@@ -1,81 +1,96 @@
 const mongoose = require('mongoose');
 mongoose.connect('mongodb://localhost/inventory');
 
+var sessionSchema = mongoose.Schema({
+  session: String,
+  userId: String
+});
+
+var userSchema = mongoose.Schema({
+  username: { type: String, unique: true },
+  salt: String,
+  password: String
+});
+
 var itemSchema = mongoose.Schema({
   name: String,
   description: String,
   quantity: Number,
-  category: String
+  category: String,
+  userId: String
 });
 
 var categorySchema = mongoose.Schema({
-  name: { type: String, unique: true, required: true},
-  itemCount: { type: Number, default: 0 }
+  name: { type: String, required: true},
+  itemCount: { type: Number, default: 0 },
+  userId: String
 })
 
 var Item = mongoose.model('Item', itemSchema);
 var Category = mongoose.model('Category', categorySchema);
+var User = mongoose.model('User', userSchema);
+var Session = mongoose.model('Session', sessionSchema);
 
-var readCats = () => {
-  return Category.find();
+var readCats = (userId) => {
+  return Category.find({ userId });
 }
 
-var createCat = (cat) => {
+var createCat = (cat, userId) => {
+  cat.userId = userId;
   var newCat = new Category(cat)
   return newCat.save();
 }
 
-var readItems = (query) => {
+var readItems = (query, userId) => {
   var category = query.category;
   var id = query.id;
-  console.log()
   var query = {};
   if (category !== 'all') {
     query.category = category;
   }
   if (id) {
-    return Item.findOne({ _id: id });
+    return Item.findOne({ _id: id, userId });
   }
+  query.userId = userId;
   return Item.find(query);
 }
 
-var createItem = (item) => {
+var createItem = (item, userId) => {
+  item.userId = userId;
   var newItem = new Item(item);
-  return Category.findOne({ name: item.category })
+  return Category.findOne({ name: item.category, userId })
     .then(cat => {
-      console.log(cat);
       cat.itemCount++
-      return Category.updateOne({ name: item.category }, { itemCount: cat.itemCount })
+      return Category.updateOne({ name: item.category, userId }, { itemCount: cat.itemCount })
     })
     .then(() => {
       return newItem.save();
     })
-
-  // return newItem.save();
 }
 
-var remove = (record) => {
+var remove = (record, userId) => {
   if (record.type === 'item') {
-    return Item.findOne({ _id: record.data._id })
+    return Item.findOne({ _id: record.data._id, userId})
       .then(item => {
-        return Category.findOne({ name: item.category });
+        return Category.findOne({ name: item.category, userId});
       })
       .then((category) => {
         category.itemCount--;
-        return Category.updateOne({ name: category.name }, category);
+        return Category.updateOne({ name: category.name, userId }, category);
       })
       .then(() => {
-        return Item.deleteMany({ _id: record.data._id });
+        return Item.deleteMany({ _id: record.data._id, userId});
       });
   } else if (record.type === 'cat') {
-    return Category.remove({ name: record.data.name })
+    return Category.remove({ name: record.data.name, userId })
       .then(() => {
-        return Item.deleteMany({ category: record.data.name });
+        return Item.deleteMany({ category: record.data.name, userId });
       })
   }
 }
 
-var updateItem = (updateItem) => {
+var updateItem = (updateItem, userId) => {
+  updateItem.item.userId = userId
   return Item.findOne(updateItem.item)
     .then(item => {
       if (updateItem.operation === 'add') {
@@ -89,7 +104,7 @@ var updateItem = (updateItem) => {
     })
 }
 
-var search = (query) => {
+var search = (query, userId) => {
   var regex = new RegExp(query.search);
   return Item.find({
     $or:[
@@ -111,8 +126,31 @@ var search = (query) => {
               $options: 'i'
             }
           }
-        ]
+        ],
+        userId
     });
+}
+
+var readSession = (session) => {
+  return Session.findOne({ session });
+}
+
+var readUser = (username) => {
+  return User.findOne({ username });
+}
+
+var createUser = (username, salt, password) => {
+  var newUser = new User({ username, salt, password });
+  return newUser.save();
+}
+
+var saveSession = (session, userId) => {
+  var newSession = new Session({ session, userId });
+  return newSession.save();
+}
+
+var destroySession = (session) => {
+  return Session.deleteOne({ session });
 }
 
 module.exports = {
@@ -122,5 +160,10 @@ module.exports = {
   readItems,
   createItem,
   updateItem,
-  search
+  search,
+  readSession,
+  readUser,
+  createUser,
+  saveSession,
+  destroySession
 }
